refactor: satisfy release source-size policy
