+++ tests/router_e2e/cases_regressions.rs
@@ -989,25 +989,3 @@ async fn a_forced_call_on_server_tools_only_fails_fast_on_every_surface() {
     }
     assert!(router.requests.lock().expect("stub requests").is_empty());
 }
-
-/// A capped server-tool search without a forced choice keeps working, so the
-/// guard above narrows nothing that previously succeeded.
-#[tokio::test]
-async fn an_uncoerced_server_tool_search_still_reaches_codex() {
-    let router = TestRouter::start(UpstreamProvider::Codex).await;
-    let response = router
-        .post(
-            "/api/services/anthropic/v1/messages",
-            &json!({
-                "model":"gpt-5",
-                "max_tokens":256,
-                "messages":[{"role":"user","content":"research Rust"}],
-                "tools":[{"type":"web_search_20250305","name":"web_search","max_uses":1}],
-                "tool_choice":{"type":"auto"}
-            }),
-        )
-        .send()
-        .await
-        .expect("server-tool response");
-    assert_eq!(response.status(), StatusCode::OK);
-}
